Extracted AddressInput component from ToolBar
