+++ desktop-app-rewrite/package.json
@@ -111,7 +111,7 @@
     "electron-debug": "^3.2.0",
     "electron-log": "^4.4.8",
     "electron-store": "^8.0.2",
-    "electron-updater": "^5.2.1",
+    "electron-updater": "^5.2.3",
     "emitter": "^0.0.5",
     "os": "^0.1.2",
     "postcss": "^8.4.14",
