+++ desktop-app-rewrite/src/renderer/components/ToolBar/NavigationControls.tsx
