+++ desktop-app-rewrite/src/renderer/components/ToolBar/index.tsx
@@ -1,0 +1,46 @@
+import { useDispatch, useSelector } from 'react-redux';
+import {
+  selectIsInspecting,
+  selectRotate,
+  setIsInspecting,
+  setRotate,
+} from 'renderer/store/features/renderer';
+import { Icon } from '@iconify/react';
+import NavigationControls from './NavigationControls';
+import Menu from './Menu';
+import Button from '../Button';
+import AddressBar from './AddressBar';
+
+const ToolBar = () => {
+  const rotateDevice = useSelector(selectRotate);
+  const isInspecting = useSelector(selectIsInspecting);
+  const dispatch = useDispatch();
+
+  return (
+    <div className="flex items-center justify-between gap-2">
+      <NavigationControls />
+      <AddressBar />
+      <Button
+        onClick={() => dispatch(setRotate(!rotateDevice))}
+        isActive={rotateDevice}
+      >
+        <Icon
+          icon={
+            rotateDevice
+              ? 'mdi:phone-rotate-portrait'
+              : 'mdi:phone-rotate-landscape'
+          }
+        />
+      </Button>
+      <Button
+        onClick={() => dispatch(setIsInspecting(!isInspecting))}
+        isActive={isInspecting}
+      >
+        <Icon icon="lucide:inspect" />
+      </Button>
+      <Menu />
+    </div>
+  );
+};
+
+export default ToolBar;
